resolved bot comments
